requirements.txt
django-allauth
python-dotenv
qrcode[pil]